Getting Started with OpenAI Whisper API/Configuring Environment/requirements.txt
Flask==2.2.5
gunicorn==22.0.0
openai>=1.57.2
python-dotenv>=1.0.1
requests>=2.31.0
